+++ src/modules/auth/AuthError.ts
